removed functions from main
